+++ CHANGELOG.md
@@ -1,2 +1,6 @@
+# 2.0.0:
+  - Split DI class in two classes: Factory and Value
+  - Allow returning multiple DI instances from middlewares and handlers
+
 # 1.0.0:
   - First release